Remove op->cache_id
Since the opid structure was increased to 128 bit, the need for the
indirection through cache_id is no longer needed.

Now the op structure pointer is typically (but not required) stored
directly in the opid.

+++ resource.h
@@ -39,8 +39,8 @@
                          __ret_type *__ae_retval,                          \
                          ##__fargs)
 
-ae_op_id_t ae_id_gen(int resource_id, uint64_t ptr);
-uint64_t ae_id_lookup(ae_op_id_t id, int *resource_id);
+ae_op_id_t ae_id_gen(int resource_id, intptr_t ptr);
+intptr_t ae_id_lookup(ae_op_id_t id, int *resource_id);
 
 /* The resource structure is defined by a given resource, and registered
  * to the aesop management code during resource initialization.
